python/agents/nutrigenomics-rag/pyproject.toml
[project]
name = "nutrigenomics-rag"
version = "0.1.0"
description = "Vertex AI RAG agent that links genomic variants to personalised nutrition, supplementation, exercise, and longevity guidance."
authors = [
    { name = "Google Sample Contributors" },
]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.8.1"
tabulate = "^0.9.0"
google-auth = "^2.36.0"
requests = "^2.32.3"
google-adk = "^1.10.0"
google-cloud-aiplatform = { extras = [
    "adk",
    "agent-engines",
], version = "^1.108.0" }
llama-index = "^0.12"
python-dotenv = "^1.0.1"

[tool.poetry.group.streamlit_runner.dependencies]
streamlit = "^1.41.1"
langchain = "^0.3.12"
langchain-community = "^0.3.12"
sentence-transformers = "^3.2.1"
faiss-cpu = "^1.8.0.post1"
pypdf = "^5.1.0"
openai = "^1.58.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
scikit-learn = "^1.6.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
